hw4 fix
